lag egen utils mappe
Mappen skal ha utilities som kan enkelt deles på tvers av løsningen
