+++ src/components/beregningOgSimulering/beregning/Beregning.tsx
@@ -24,7 +24,6 @@ import {
 import ToKolonner from '~components/toKolonner/ToKolonner';
 import { kanSimuleres } from '~features/behandling/behandlingUtils';
 import fradragstypeMessages from '~features/fradrag/fradragstyper-nb';
-import { hentBosituasjongrunnlag } from '~features/revurdering/revurderingUtils';
 import * as sakSlice from '~features/saksoversikt/sak.slice';
 import * as DateUtils from '~lib/dateUtils';
 import { useI18n } from '~lib/hooks';
@@ -35,6 +34,7 @@ import { useAppDispatch, useAppSelector } from '~redux/Store';
 import { Behandlingsstatus } from '~types/Behandling';
 import { Beregning } from '~types/Beregning';
 import { Fradrag, Fradragstype, FradragTilhører } from '~types/Fradrag';
+import { hentBosituasjongrunnlag } from '~Utils/revurdering/revurderingUtils';
 
 import sharedI18n from '../../../pages/saksbehandling/søknadsbehandling/sharedI18n-nb';
 import { VilkårsvurderingBaseProps } from '../../../pages/saksbehandling/søknadsbehandling/types';
