+++ src/components/oppsummering/vilkårsOppsummering/VilkårsOppsummering.tsx
@@ -1,7 +1,6 @@
 import { Systemtittel } from 'nav-frontend-typografi';
 import React from 'react';
 
-import { hentBosituasjongrunnlag } from '~features/revurdering/revurderingUtils';
 import { mapToVilkårsinformasjon, Vilkårsinformasjon } from '~features/saksoversikt/utils';
 import { useI18n } from '~lib/hooks';
 import { Behandlingsstatus } from '~types/Behandling';
@@ -9,6 +8,7 @@ import { Behandlingsinformasjon } from '~types/Behandlingsinformasjon';
 import { GrunnlagsdataOgVilkårsvurderinger } from '~types/grunnlagsdataOgVilkårsvurderinger/grunnlagsdataOgVilkårsvurderinger';
 import { SøknadInnhold } from '~types/Søknad';
 import { Vilkårtype } from '~types/Vilkårsvurdering';
+import { hentBosituasjongrunnlag } from '~Utils/revurdering/revurderingUtils';
 
 import { FastOppholdVilkårsblokk } from './faktablokk/faktablokker/FastOppholdFaktablokk';
 import { FlyktningVilkårsblokk } from './faktablokk/faktablokker/FlyktningFaktablokk';
