+++ src/pages/saksbehandling/revurdering/endringAvFradrag/EndringAvFradrag.tsx
@@ -18,7 +18,6 @@ import RevurderingskallFeilet from '~components/revurdering/revurderingskallFeil
 import ToKolonner from '~components/toKolonner/ToKolonner';
 import fradragstypeMessages from '~features/fradrag/fradragstyper-nb';
 import { lagreFradragsgrunnlag } from '~features/revurdering/revurderingActions';
-import { hentBosituasjongrunnlag } from '~features/revurdering/revurderingUtils';
 import sharedMessages from '~features/revurdering/sharedMessages-nb';
 import * as DateUtils from '~lib/dateUtils';
 import { customFormikSubmit } from '~lib/formikUtils';
@@ -29,6 +28,7 @@ import { useAppDispatch } from '~redux/Store';
 import { Fradrag, FradragTilhører } from '~types/Fradrag';
 import { GrunnlagsdataOgVilkårsvurderinger } from '~types/grunnlagsdataOgVilkårsvurderinger/grunnlagsdataOgVilkårsvurderinger';
 import { Revurdering } from '~types/Revurdering';
+import { hentBosituasjongrunnlag } from '~Utils/revurdering/revurderingUtils';
 
 import uføreMessages from '../../søknadsbehandling/uførhet/uførhet-nb';
 import { RevurderingBunnknapper } from '../bunnknapper/RevurderingBunnknapper';
