+++ src/pages/søknad/steg/formue/søkersFormue/DinFormue.tsx
@@ -7,12 +7,13 @@ import { JaNeiSpørsmål } from '~/components/formElements/FormElements';
 import søknadSlice, { SøknadState } from '~/features/søknad/søknad.slice';
 import { useI18n } from '~lib/hooks';
 import { keyOf } from '~lib/types';
-import { formikErrorsTilFeiloppsummering, formikErrorsHarFeil, formueValideringSchema } from '~lib/validering';
+import { formikErrorsTilFeiloppsummering, formikErrorsHarFeil } from '~lib/validering';
 import { useAppSelector, useAppDispatch } from '~redux/Store';
 
 import Bunnknapper from '../../../bunnknapper/Bunnknapper';
 import sharedStyles from '../../../steg-shared.module.less';
 import sharedI18n from '../../steg-shared-i18n';
+import { formueValideringSchema } from '../formueSøknadUtils';
 import KjøretøyInputFelter from '../KjøretøyInputFelter';
 
 import messages from './dinformue-nb';
